+++ demo/gen-boms.sh
@@ -1,14 +1,13 @@
 #!/bin/sh
-
-## purpose: generate example reguslts from the demos
-
 set -e
+
+## purpose: generate example results from the demos
 
 THIS_DIR="$(dirname "$0")"
 BIN_CDX_N="$(realpath "$THIS_DIR/../bin/cyclonedx-npm-cli.js")"
 export npm_execpath="${npm_execpath:-$(which npm)}"
 
-for package in $THIS_DIR/*/project/package.json
+for package in "$THIS_DIR"/*/project/package.json
 do
   echo ">>> $package"
   project="$(dirname "$package")"
@@ -22,12 +21,23 @@ do
   do
     for spec in '1.2' '1.3' '1.4'
     do
-      echo ">>> $result_dir $spec $format"
+      echo ">>> $result_dir $spec $format bare"
+      mkdir -p "$result_dir/bare"
       node -- "$BIN_CDX_N" \
       --spec-version "$spec" \
       --output-reproducible \
       --output-format "$format" \
-      --output-file "$result_dir/bom.$spec.$format" \
+      --output-file "$result_dir/bare/bom.$spec.$format" \
+      "$package"
+
+      echo ">>> $result_dir $spec $format flat"
+      mkdir -p "$result_dir/flat"
+      node -- "$BIN_CDX_N" \
+      --flatten-components \
+      --spec-version "$spec" \
+      --output-reproducible \
+      --output-format "$format" \
+      --output-file "$result_dir/flat/bom.$spec.$format" \
       "$package"
     done
   done